Import patches-unapplied version 0.9.10-1ubuntu1 to ubuntu/hardy
Imported using git-ubuntu import.

Changelog parent: 59e131882e0714d1eec80251f280ae960a848ee9

New changelog entries:
  [ Daniel T Chen ]
  * Merge from Debian unstable:
  * Patches
    + 0050-Reduce-RT-highprio-log-to-info:
    + 0051-Reduce-pa_pid_file_create-Daemon-already-running-log-spam:
      Retain to prevent innocuous messages from spamming user log;
    - 0052-Add-extra-checks-to-padsp-debug-calls:
    - 0053-Handle-channel-map-failure:
    - 0054-Invoke-pa_ltdl_done-conditionally:
    - 0055-fix_record_stream_moved:  Merged in new upstream version.
  * Packaging
    + control:  Adhere to DebianMaintainerField spec, and add a
      Vcs-Bzr entry;
    + control:
    + pulseaudio-module-jack.install:  Don't generate or install the
      jackd module;
    + pulseaudio.init:
    + rules:  Use multiuser semantics.
  [ Luke Yelavich ]
  * FFe granted. (LP: #211592)
  [ CJ van den Berg ]
  * New Upstream Version
  * debian/patches: Drop patches merged upstream
    - 0002-Double-esound-maximum-sample-size.patch
    - 0003-fix-uploading-of-samples-into-PA.-Problem-discovered.patch
    - 0004-make-sure-to-create-.pulse-before-using-any-config.patch
    - 0005-Don-t-add-protocol-fields-introduced-in-version-12-w.patch
    - 0006-Implement-opcodes-added-in-version-12-in-the-tunnel.patch
    - 0007-Change-config-to-not-fail-if-loading-of-esd-gconf-o.patch
    - 0008-Fix-compilation-on-non-linux-platforms.patch
  * debian/patches: Add patch to fix RPATHs.
    + 0002-Fix-library-search-path-to-include-lib-and-usr-lib.patch
  * debian/control:
    + Make libpulsecore5-dbg depend on libpulsecore5, instead of pulseaudio.
    + Make libpulse-browse0-dbg depend on libpulse-browse0.
  * Update Standards-Version to 3.7.3 (no changes required).
  [ Petter Reinholdtsen ]
  * debian/pulseaudio.init: Fix problems with LSB header in init.d script.
      (Closes: #470934)

+++ src/tests/smoother-test.c
@@ -1,4 +1,4 @@
-/* $Id: smoother-test.c 1971 2007-10-28 19:13:50Z lennart $ */
+/* $Id: smoother-test.c 2191 2008-03-29 23:09:48Z lennart $ */
 
 /***
   This file is part of PulseAudio.
@@ -71,7 +71,7 @@ int main(int argc, char*argv[]) {
             u += 2;
         }
 
-        printf("%llu\t%llu\n", x/PA_USEC_PER_MSEC, pa_smoother_get(s, x)/PA_USEC_PER_MSEC);
+        printf("%llu\t%llu\n", (unsigned long long) (x/PA_USEC_PER_MSEC), (unsigned long long) (pa_smoother_get(s, x)/PA_USEC_PER_MSEC));
     }
 
     pa_smoother_free(s);
